refactor: 식당 상세 정보 세부 컴포넌트 작명 수정
