+++ client/src/components/RestaurantDetailModal/RestaurantDetailModalTitle/index.tsx
@@ -1,0 +1,21 @@
+import { CategoryBox, ModalTitleBox, ModalTitleLayout, NameBox, RatingBox } from './styles';
+
+interface PropsType {
+  name: string;
+  category: string;
+  rating: number;
+}
+
+export function RestaurantDetailModalBody({ name, category, rating = 0 }: PropsType) {
+  return (
+    <ModalTitleLayout>
+      <ModalTitleBox>
+        <NameBox>
+          <p>{name}</p>
+        </NameBox>
+        <CategoryBox>{category}</CategoryBox>
+        <RatingBox>{!rating ? '평점 정보 없음' : `평점: ${rating}`}</RatingBox>
+      </ModalTitleBox>
+    </ModalTitleLayout>
+  );
+}
